refactor(backend): move library sync out of auth.routes into infrastructure
Library sync for Steam/Epic/GOG lived as exported functions in
auth.routes.ts, forcing group.routes to dynamically import a presentation
module 'to avoid circular deps' and inverting the dependency direction
(domain coordinator implemented by a route file).

Move the three sync functions to infrastructure/library/library-sync.ts
and add createLibrarySyncCoordinator()/syncGroupLibraries() factories so
any caller can reuse them without a presentation-layer dependency.
