+++ packages/backend/src/presentation/routes/auth.routes.ts
@@ -1,10 +1,10 @@
 import crypto from 'crypto'
 import { Router, type Request, type Response } from 'express'
 import { db } from '../../infrastructure/database/connection.js'
-import { getSteamLoginUrl, verifySteamLogin, getPlayerSummary, getOwnedGames, getHeaderImageUrl } from '../../infrastructure/steam/steam-client.js'
-import { isEpicEnabled, getEpicAuthUrl, exchangeCodeForTokens, getOwnedGames as getEpicOwnedGames, normalizeGameName } from '../../infrastructure/epic/epic-client.js'
-import { isGogEnabled, getGogAuthUrl, exchangeCodeForTokens as exchangeGogCode, getOwnedGames as getGogOwnedGames, normalizeGameName as normalizeGogGameName } from '../../infrastructure/gog/gog-client.js'
-import type { NewlyAcquiredGame } from '../../domain/new-game-spotlight.js'
+import { getSteamLoginUrl, verifySteamLogin, getPlayerSummary } from '../../infrastructure/steam/steam-client.js'
+import { isEpicEnabled, getEpicAuthUrl, exchangeCodeForTokens } from '../../infrastructure/epic/epic-client.js'
+import { isGogEnabled, getGogAuthUrl, exchangeCodeForTokens as exchangeGogCode } from '../../infrastructure/gog/gog-client.js'
+import { syncUserLibrary, syncEpicLibrary, syncGogLibrary } from '../../infrastructure/library/library-sync.js'
 import { encryptToken } from '../../infrastructure/crypto/token-cipher.js'
 import { authLogger, steamLogger, epicLogger, gogLogger } from '../../infrastructure/logger/logger.js'
 import { env } from '../../config/env.js'
@@ -936,231 +936,4 @@ router.post('/gog/sync', requireAuth, async (req: Request, res: Response) => {
   }
 })
 
-// ─── Background Library Sync ────────────────────────────────────────
-
-// Background Epic library sync
-async function syncEpicLibrary(userId: string): Promise<number> {
-  const games = await getEpicOwnedGames(userId)
-  if (!games || games.length === 0) {
-    epicLogger.warn({ userId }, 'no Epic games returned or token issue')
-    return 0
-  }
-
-  const now = new Date()
-  for (const game of games) {
-    let gameId: string | null = null
-    const normalizedName = normalizeGameName(game.displayName)
-
-    // Check if this Epic game already has a platform mapping
-    const existingMapping = await db('game_platform_ids')
-      .where({ platform: 'epic', platform_game_id: game.catalogItemId })
-      .first()
-
-    if (existingMapping) {
-      gameId = existingMapping.game_id
-    } else {
-      // Try to find a canonical game with matching normalized name
-      const existingGame = await db('games')
-        .whereRaw('LOWER(REGEXP_REPLACE(canonical_name, \'[^a-zA-Z0-9\\s]\', \'\', \'g\')) = ?', [normalizedName])
-        .first()
-
-      if (existingGame) {
-        gameId = existingGame.id
-      } else {
-        const [newGame] = await db('games')
-          .insert({ canonical_name: game.displayName })
-          .returning('id')
-        gameId = newGame.id
-      }
-
-      await db('game_platform_ids').insert({
-        game_id: gameId,
-        platform: 'epic',
-        platform_game_id: game.catalogItemId,
-      })
-    }
-
-    await db('user_games')
-      .insert({
-        user_id: userId,
-        game_id: gameId,
-        platform: 'epic',
-        game_name: game.displayName,
-        synced_at: now,
-      })
-      .onConflict(['user_id', 'game_id', 'platform'])
-      .merge({
-        game_name: game.displayName,
-        synced_at: now,
-      })
-  }
-
-  epicLogger.info({ userId, gameCount: games.length }, 'Epic library synced')
-  return games.length
-}
-
-// Background Steam library sync
-async function syncUserLibrary(userId: string, steamId: string): Promise<number> {
-  const games = await getOwnedGames(steamId)
-  if (!games) {
-    await db('users').where({ id: userId }).update({ library_visible: false })
-    return 0
-  }
-
-  if (games.length === 0) {
-    steamLogger.warn({ steamId }, 'no games returned — profile may be private')
-    await db('users').where({ id: userId }).update({ library_visible: false })
-    return 0
-  }
-
-  // Detect freshly acquired games for the new-game spotlight. Snapshot the
-  // library BEFORE the upsert so we can tell which app IDs are genuinely new.
-  // The very first sync seeds the entire library, so we never spotlight on it.
-  const existingAppIds = new Set<number>(
-    await db('user_games').where({ user_id: userId, platform: 'steam' }).pluck('steam_app_id'),
-  )
-  const isFirstSync = existingAppIds.size === 0
-  const newlyAcquired: NewlyAcquiredGame[] = []
-
-  // Upsert all games
-  const now = new Date()
-  for (const game of games) {
-    // Find or create canonical game entry
-    let gameId: string | null = null
-    const existingMapping = await db('game_platform_ids')
-      .where({ platform: 'steam', platform_game_id: String(game.appid) })
-      .first()
-
-    if (existingMapping) {
-      gameId = existingMapping.game_id
-    } else {
-      // Create canonical game + platform mapping
-      const [newGame] = await db('games')
-        .insert({
-          canonical_name: game.name,
-          cover_image_url: getHeaderImageUrl(game.appid),
-        })
-        .returning('id')
-      gameId = newGame.id
-      await db('game_platform_ids').insert({
-        game_id: gameId,
-        platform: 'steam',
-        platform_game_id: String(game.appid),
-      })
-    }
-
-    await db('user_games')
-      .insert({
-        user_id: userId,
-        steam_app_id: game.appid,
-        game_id: gameId,
-        platform: 'steam',
-        game_name: game.name,
-        header_image_url: getHeaderImageUrl(game.appid),
-        playtime_forever: game.playtime_forever ?? null,
-        playtime_2weeks: game.playtime_2weeks ?? null,
-        synced_at: now,
-      })
-      .onConflict(['user_id', 'steam_app_id'])
-      .merge({
-        game_name: game.name,
-        game_id: gameId,
-        header_image_url: getHeaderImageUrl(game.appid),
-        playtime_forever: db.raw('GREATEST(EXCLUDED.playtime_forever, user_games.playtime_forever)'),
-        playtime_2weeks: game.playtime_2weeks ?? null,
-        synced_at: now,
-      })
-
-    if (!isFirstSync && gameId && !existingAppIds.has(game.appid)) {
-      newlyAcquired.push({
-        steamAppId: game.appid,
-        gameId,
-        gameName: game.name,
-        headerImageUrl: getHeaderImageUrl(game.appid),
-      })
-    }
-  }
-
-  await db('users').where({ id: userId }).update({ library_visible: true, updated_at: now })
-  steamLogger.info({ userId, steamId, gameCount: games.length }, 'library synced')
-
-  // Evaluate playtime/collection/dedication challenges after sync (non-blocking)
-  const { evaluateChallenges } = await import('../../domain/challenges/challenge-service.js')
-  evaluateChallenges(userId, ['playtime', 'dedication', 'collection']).catch(err =>
-    steamLogger.warn({ error: String(err), userId }, 'challenge evaluation after sync failed')
-  )
-
-  // Announce freshly acquired games into the member's groups (non-blocking).
-  // Dynamic import keeps this off the module-load critical path and avoids a
-  // require cycle (the spotlight module pulls in create-session, which would
-  // otherwise tangle with the auth routes at load time).
-  if (newlyAcquired.length > 0) {
-    const { processNewGameSpotlights } = await import('../../domain/new-game-spotlight.js')
-    processNewGameSpotlights(userId, newlyAcquired).catch(err =>
-      steamLogger.warn({ error: String(err), userId }, 'new game spotlight processing failed')
-    )
-  }
-
-  return games.length
-}
-
-// Background GOG library sync
-async function syncGogLibrary(userId: string): Promise<number> {
-  const games = await getGogOwnedGames(userId)
-  if (!games || games.length === 0) {
-    gogLogger.warn({ userId }, 'no GOG games returned or token issue')
-    return 0
-  }
-
-  const now = new Date()
-  for (const game of games) {
-    let gameId: string | null = null
-    const normalizedName = normalizeGogGameName(game.title)
-
-    const existingMapping = await db('game_platform_ids')
-      .where({ platform: 'gog', platform_game_id: String(game.id) })
-      .first()
-
-    if (existingMapping) {
-      gameId = existingMapping.game_id
-    } else {
-      const existingGame = await db('games')
-        .whereRaw('LOWER(REGEXP_REPLACE(canonical_name, \'[^a-zA-Z0-9\\s]\', \'\', \'g\')) = ?', [normalizedName])
-        .first()
-
-      if (existingGame) {
-        gameId = existingGame.id
-      } else {
-        const [newGame] = await db('games')
-          .insert({ canonical_name: game.title })
-          .returning('id')
-        gameId = newGame.id
-      }
-
-      await db('game_platform_ids').insert({
-        game_id: gameId,
-        platform: 'gog',
-        platform_game_id: String(game.id),
-      })
-    }
-
-    await db('user_games')
-      .insert({
-        user_id: userId,
-        game_id: gameId,
-        platform: 'gog',
-        game_name: game.title,
-        synced_at: now,
-      })
-      .onConflict(['user_id', 'game_id', 'platform'])
-      .merge({
-        game_name: game.title,
-        synced_at: now,
-      })
-  }
-
-  gogLogger.info({ userId, gameCount: games.length }, 'GOG library synced')
-  return games.length
-}
-
-export { router as authRoutes, syncUserLibrary, syncEpicLibrary, syncGogLibrary }
+export { router as authRoutes }
